提交代码

+++ newsManager/src/main/java/com/yulu/mangger/controller/CollectController.java
@@ -1,6 +1,7 @@
 package com.yulu.mangger.controller;
 
 import com.alibaba.fastjson.JSONObject;
+import com.yulu.mangger.ErrorCode;
 import com.yulu.mangger.bean.*;
 import com.yulu.mangger.service.CollectsService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 @Controller
 // 定义该Controller的根访问路径 /collect
@@ -20,43 +22,36 @@ public class CollectController {
 	@Qualifier("collectsService")
 	private CollectsService collectsService;
 
-
 	// 收藏新闻
 	@RequestMapping("/collect")
-	public String collect(HttpServletRequest request, Integer newsid,
-						  Integer userid) throws Exception {
-		Collects collects = new Collects();
-		collects.setUserid(userid);
-		collects.setNewsid(newsid);
-		collects.setIsdelete(0);
-		collectsService.add_do(collects);
-		return "forward:user_collect?userid=" + userid;
-	}
-
-	// 收藏新闻
-	@RequestMapping("/collect2")
-	public void collect(Integer newsid, Integer userid,
+	public void collect(HttpSession session, Integer newsid,
 						HttpServletResponse response) throws Exception {
+		Integer userid = (Integer) session.getAttribute("userid");
 		ResultBean<Collects> mResultBean = new ResultBean<Collects>();
-		Collects collectsOld = collectsService.findCollects(newsid,userid);
-		if (collectsOld != null) {
-			mResultBean.setMsg("已经保存");
-			mResultBean.setCode(1);
-			mResultBean.setData(collectsOld);
-		}else {
-			mResultBean.setMsg("保存成功");
-			mResultBean.setCode(0);
-			Collects collects = new Collects();
-			collects.setUserid(userid);
-			collects.setNewsid(newsid);
-			collects.setIsdelete(0);
-			collectsService.add_do(collects);
-		}
 		response.setContentType("application/json; charset=utf-8");
+		if(userid!=null) {
+			Collects collectsOld = collectsService.findCollects(newsid, userid);
+			if (collectsOld != null) {
+				mResultBean.setMsg(ErrorCode.getMsg(ErrorCode.COLLECT_SAVE_ALREADY));
+				mResultBean.setCode(ErrorCode.COLLECT_SAVE_ALREADY);
+				mResultBean.setData(collectsOld);
+			} else {
+				mResultBean.setMsg("保存成功");
+				mResultBean.setCode(ErrorCode.SUCCESS);
+				Collects collects = new Collects();
+				collects.setUserid(userid);
+				collects.setNewsid(newsid);
+				collects.setIsdelete(0);
+				collectsService.add_do(collects);
+			}
+		}else{
+			mResultBean.setMsg(ErrorCode.getMsg(ErrorCode.LOGIN_NOT_LOGIN));
+			mResultBean.setCode(ErrorCode.LOGIN_NOT_LOGIN);
+		}
 		response.getWriter().println(JSONObject.toJSONString(mResultBean));
 	}
 
-	// 收藏新闻
+	// 是否收藏新闻
 	@RequestMapping("/isCollect")
 	public void isCollect(Integer newsid, Integer userid,
 						  HttpServletResponse response) throws Exception {
@@ -76,13 +71,23 @@ public class CollectController {
 
 	// 取消收藏
 	@RequestMapping("/delete_do")
-	public void delete_do( Integer id,
-							Integer userid,HttpServletResponse response) throws Exception {
-		collectsService.delete_do(id);
+	public void delete_do( HttpSession session,Integer id,HttpServletResponse response) throws Exception {
 		ResultBean<Collects> mResultBean = new ResultBean<Collects>();
 		response.setContentType("application/json; charset=utf-8");
-		mResultBean.setMsg("取消收藏成功");
-		mResultBean.setCode(0);
+
+		Integer userid = (Integer) session.getAttribute("userid");
+		if(userid!=null) {
+			Collects collectsOld = collectsService.findCollects(id, userid);
+			if (collectsOld != null) {
+				collectsService.delete_do(id);
+				mResultBean.setMsg("取消收藏成功");
+				mResultBean.setCode(ErrorCode.SUCCESS);
+			}else{
+				mResultBean.setMsg(ErrorCode.getMsg(ErrorCode.COLLECT_NOT_COLLECT));
+				mResultBean.setCode(ErrorCode.COLLECT_NOT_COLLECT);
+			}
+		}
+
 		response.getWriter().println(JSONObject.toJSONString(mResultBean));
 	}
